e2e bug fix update and refactor
