+++ apps/festival/festival-e2e/src/integration/profile/profile-page.ts
@@ -2,22 +2,17 @@ import { loginWithRandomUser, clearBrowserAuth, loginWithUID, loginWithEmailAndP
 
 describe('Basic Landing Page, Login and Profile Page Tests', () => {
 
-  before(() => {
-    cy.visit('/')
-  })
   beforeEach(() => {
     cy.visit('/')
     clearBrowserAuth()
     cy.visit('/')
-    cy.contains('Accept cookies').click();
   })
   it('should load user and log into profile page', () => {
     // loginWithRandomUser().logSubject(); => does not work because of missing sign_in_provider in token
     // loginWithUID('1QErdmriBqaQyobxftpWyK1ErB73').logSubject(); => same
     // loginWithRandomUser('emailAndPassword').logSubject(); => do work but always fetch a blockframesAdmin user
-    loginWithEmailAndPassword('dev+josefa-aoi@blockframes.io').logSubject(); // => do work
+    loginWithEmailAndPassword('dev+scarlett-qra@blockframes.io').logSubject(); // => do work
     cy.visit('c/o/account/profile/view/settings');
-    cy.pause();
     cy.contains('Contact Information').should('exist');
   })
   it('should now show Accept Cookies after being accepted', () => {
@@ -27,9 +22,6 @@ describe('Basic Landing Page, Login and Profile Page Tests', () => {
   })
   it('should show the Accept Cookies banner', () => {
     cy.contains('Accept cookies').should('exist');
-  })
-  afterEach(async () => {
-    clearBrowserAuth()
   })
 
 })
